style: :art: change some names of values
