+++ logic/Preparation/Interface/ICharacter.cs
@@ -13,7 +13,7 @@ namespace Preparation.Interface
         public BulletType BulletOfPlayer { get; set; }
         public CharacterType CharacterType { get; }
         public ActiveSkill FindActiveSkill(ActiveSkillType activeSkillType);
-        public IntNumUpdateByCD BulletNum { get; }
+        public IntNumUpdateEachCD BulletNum { get; }
         public long SetPlayerState(RunningStateType running, PlayerStateType value = PlayerStateType.Null, IGameObj? obj = null);
         public bool ResetPlayerState(long state, RunningStateType running = RunningStateType.Null, PlayerStateType value = PlayerStateType.Null, IGameObj? obj = null);
